Optimised shape creation functions and improved actor memory handling
Create shape function from dynamic & static actor classes now only passes the current shape's color pointer to the renderer, instead of looping unnecessarily through previous shape's pointers.

Above fix exposed bug where the addresses of the color vector elements (a vector of values) were passed to the renderer, which led to dangling pointers when the vector's size was increased (and its memory re-assigned). Fixed by making the color vector a vector of pointers, each of which is passed directly to the renderer.