Updated pointers to unique pointers
Changed color vector to vector of smart pointers. May need to re-think if this is a good design choice due to the rule of 5 - this seems like a lot of code for something that could use value semantics. Also need to look into how to handle actor pointer with copy/move constructors/operators.