Improved DynamicActor's mass/inertia tensor calculations and simplified Actor&Dynamic/Static Actor's constructors and CreateShape functions
Updated mass and inertia tensor function in DynamicActor class to correctly handle actors with shapes of different densities.

Changed Actor constructor so that it accepts an instance of PxRigidActor to assign to "actor" - derived wrapper classes can therefore pass instantiated derived classes of "PxRigidActor".

Changed CreateShape function to have less repeated code. No longer a pure virtual function, instead the shared implementation is written in Actor.h, and a virtual ChangeShapeHelper
function allows the derrived classes to add more functionality to the end of the CreateShape function. This allows the PxRigidDynamic class to add on implementation relating to
shape densities. Note that this functionality could be reversed, so that the derived classes override a pure virtual, and call a base class helper to implement shared functionality.

TODO - need to look at switching UserData instances in Actor class to unique_ptr, and passing an observing raw pointer to the Shape instance.
Also need to create an interface in Actor.h to change local pos of shape and then update inertia tensor, as currently the raw shape pointer is being exposed to do this,
meaning the inertia tensor is not updated. See shape getter for details.