Updated clients to use new SetShapePos function
Updated clients to use new SetShapePos function. Note that this has broken the scene's physics (can't get any height on the ball without scooping it).
However, it was already using incorrect physics (materials are still all wrong), so this is not a problem - better to get the engine working correctly before changing scene.

TODO - move on to looking at joint classes